Split RiscVVersionInfo to have child type RiscVExtensionInfo
